(SLV-375) Use git ref of golja-influxdb
The latest current release of the `golja-influxdb` module does not
support puppet 5.  However, the module is still under development and
the git repo master branch does support puppet 5.

This commit updates the Puppetfile to use a git ref of the influxdb
module that supports puppet 5.  This reference should be updated when a
new release of this module is published to the forge.