(SLV-422) Add spec test for module list
This commit adds a spec test that ensures that the installed module tree
has no invalid entries.  Rake tasks are added to install the modules and
to run the spec tests.

A .gitignore file is added to prevent any of the artifacts from this
process being added to version control.