(SLV-422) Add puppet-lint config
This commit adds a configuration file for puppet-lint.  Its primary
purpose is to ignore anything in the `modules` directory that is pulled
down for testing.

Most of the file consists of skipping validations that currently
generate errors.  These `--no-*` lines should be removed when the code
has been corrected to enable these checks to pass.

Ref: [SLV-674](https://tickets.puppetlabs.com/browse/SLV-674)